+++ packages/framework/src/Actions/PageModelConstructor.php
@@ -2,8 +2,12 @@
 
 namespace Hyde\Framework\Actions;
 
+use Hyde\Framework\Contracts\AbstractMarkdownPage;
 use Hyde\Framework\Contracts\AbstractPage;
+use Hyde\Framework\Hyde;
+use Hyde\Framework\Models\Pages\BladePage;
 use Hyde\Framework\Models\Pages\DocumentationPage;
+use Illuminate\Support\Str;
 
 /**
  * Dynamically constructs data for a page model.
@@ -25,15 +29,50 @@ class PageModelConstructor
 
     protected function constructDynamicData(): void
     {
-        $this->page->title = SourceFileParser::findTitleForPage($this->page, $this->page->identifier);
+        $this->page->title = self::findTitleForPage($this->page, $this->page->identifier);
 
         if ($this->page instanceof DocumentationPage) {
-            $this->page->category = SourceFileParser::getDocumentationPageCategory($this->page, $this->page->identifier);
+            $this->page->category = self::getDocumentationPageCategory($this->page, $this->page->identifier);
         }
     }
 
     protected function get(): AbstractPage
     {
         return $this->page;
+    }
+
+    public static function getDocumentationPageCategory(DocumentationPage $page, string $slug): ?string
+    {
+        // If the documentation page is in a subdirectory,
+        // then we can use that as the category name.
+        // Otherwise, we look in the front matter.
+
+        return str_contains($slug, '/')
+            ? Str::before($slug, '/')
+            : $page->matter('category');
+    }
+
+    public static function findTitleForPage(BladePage|AbstractMarkdownPage $page, string $slug): string
+    {
+        if ($page instanceof BladePage) {
+            return Hyde::makeTitle($slug);
+        }
+
+        if ($page->matter('title')) {
+            return $page->matter('title');
+        }
+
+        return static::findTitleFromMarkdownHeadings($page) ?? Hyde::makeTitle($slug);
+    }
+
+    public static function findTitleFromMarkdownHeadings(AbstractMarkdownPage $page): ?string
+    {
+        foreach ($page->markdown()->toArray() as $line) {
+            if (str_starts_with($line, '# ')) {
+                return trim(substr($line, 2), ' ');
+            }
+        }
+
+        return null;
     }
 }
